Improved method for running in a debugger
A debug function was added to the MunitTest struct as well as the
test_definition struct. When RUN_IN_DEBUGGER is set to true, the program
runs these debug functions instead of the tests. The debug functions
must be written such that they can run in a debugger (e.g. they can't
fork the process, as is done in multiple test functions.)